[ruby/digest] Fix Digest::SHA1#update with large input
Digest::SHA1#update fails when a very large String is passed in a
single call.

Passing 2**29 bytes (512 MB) or more at once does not update the
message length counter correctly, which results in producing an
incorrect output.

    $ ruby -rdigest -e'd=Digest::SHA1.new; d<<"a"*(512*1024*1024); puts d.hexdigest'
    https://github.com/ruby/digest/commit/40cd9c4b14e7
    $ ruby -rdigest -e'd=Digest::SHA1.new; 512.times{d<<"a"*(1024*1024)}; puts d.hexdigest'
    https://github.com/ruby/digest/commit/0ea59bfe8787
    $ ruby -e'print "a"*(512*1024*1024)'|sha1sum
    https://github.com/ruby/digest/commit/0ea59bfe8787  -

Passing 2**32 bytes or more causes an infinite loop because the loop
counter is too small.

https://github.com/ruby/digest/commit/1bf687bf75

Co-authored-by: Nobuyoshi Nakada <nobu@ruby-lang.org>